add trigger_recommend
